add playerSckin

+++ iichanTouhou/Objects/Bullets/Behavior/Collision/StatChanger.cs
@@ -3,12 +3,12 @@
     class StatChanger
     {
         private readonly int _valueOfChangeOfXp;
-        private readonly int _valueOfChangeOfPower;
+        private readonly float _valueOfChangeOfPower;
         private readonly int _valueOfChangeOfScore;
         private readonly int _valueOfChangeOfcountOfLives;
 
 
-        public StatChanger(int valueOfChangeOfXp, int valueOfChangeOfPower, int valueOfChangeOfScore, int valueOfChangeOfcountOfLives)
+        public StatChanger(int valueOfChangeOfXp, float valueOfChangeOfPower, int valueOfChangeOfScore, int valueOfChangeOfcountOfLives)
         {
             _valueOfChangeOfXp = valueOfChangeOfXp;
             _valueOfChangeOfPower = valueOfChangeOfPower;
@@ -16,7 +16,7 @@
             _valueOfChangeOfcountOfLives = valueOfChangeOfcountOfLives;
         }
 
-        public StatChanger(int valueOfChangeOfXp, int valueOfChangeOfPower, int valueOfChangeOfScore)
+        public StatChanger(int valueOfChangeOfXp, float valueOfChangeOfPower, int valueOfChangeOfScore)
             :this(valueOfChangeOfXp,valueOfChangeOfPower,valueOfChangeOfScore,0)
         {
             _valueOfChangeOfXp = valueOfChangeOfXp;
